Move actions, reducers, types to store directory
