+++ src/store/index.ts
@@ -1,0 +1,16 @@
+import {
+    combineReducers,
+    Dispatch,
+    Reducer,
+    Action,
+    AnyAction,
+} from 'redux';
+import terminalReducer from './terminal/reducer';
+
+const rootReducer = combineReducers({
+    terminal: terminalReducer,
+});
+
+export default rootReducer;
+
+// export { default as rootReducer } from './root';
