+++ src/store/root.ts
@@ -1,0 +1,8 @@
+// import { combineReducers } from 'redux';
+// import terminalReducer from './terminal/reducer';
+//
+// const rootReducer = combineReducers({
+//     terminal: terminalReducer,
+// });
+//
+// export default rootReducer;
